+++ crates/vmi-core/src/core/mod.rs
@@ -1,6 +1,5 @@
 mod access_context;
 mod address_context;
-mod hex;
 mod info;
 mod macros;
 mod memory_access;
@@ -11,7 +10,6 @@ use self::macros::impl_ops;
 pub use self::{
     access_context::{AccessContext, Gfn, Pa, TranslationMechanism, Va, VmiVa},
     address_context::AddressContext,
-    hex::Hex,
     info::VmiInfo,
     memory_access::{MemoryAccess, MemoryAccessOptions},
     vcpu_id::VcpuId,