+++ crates/vmi-core/src/lib.rs
@@ -20,7 +20,7 @@ use zerocopy::{FromBytes, Immutable, IntoBytes};
 pub use self::{
     arch::{Architecture, Registers},
     core::{
-        AccessContext, AddressContext, Gfn, Hex, MemoryAccess, MemoryAccessOptions, Pa,
+        AccessContext, AddressContext, Gfn, MemoryAccess, MemoryAccessOptions, Pa,
         TranslationMechanism, Va, VcpuId, View, VmiInfo, VmiVa,
     },
     ctx::{